requirements.txt
streamlit==1.43.2
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
scipy==1.11.2
scikit-image==0.21.0
